[REF] mail: move discuss core files to core folder
This prevents from having files floating in the root folder and making
the navigation between subfolders more complex.

Part of task-3265211
