+++ addons/mail/static/src/discuss/core/attachment_model_patch.js
