+++ addons/mail/static/src/discuss/core/call_settings.xml
