+++ addons/mail/static/src/discuss/core/channel_commands.js
