+++ addons/mail/static/src/discuss/core/channel_invitation.js
