+++ addons/mail/static/src/discuss/core/channel_invitation.xml
