+++ addons/mail/static/src/discuss/core/channel_member_list.js
