+++ addons/mail/static/src/discuss/core/discuss_patch.xml
