+++ addons/mail/static/src/discuss/core/suggestion_service_patch.js
