+++ addons/mail/static/src/discuss/core/thread_service_patch.js
